claudedocs/prototypes/claude_docai_sklton_files/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LLM Providers
openai==1.3.0
langchain==0.1.0
langchain-openai==0.0.2
langchain-core==0.1.0

# LLM Graph (optional for advanced workflows)
langgraph==0.0.20

# Vector Databases
pymilvus==2.3.3
faiss-cpu==1.7.4
sentence-transformers==2.2.2

# Relational Database
duckdb==0.9.2

# Caching
redis==5.0.1
hiredis==2.2.3

# Utilities
python-multipart==0.0.6
python-dotenv==1.0.0

# Optional: For production
gunicorn==21.2.0
